backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pydantic>=2.0.0
python-dateutil>=2.8.0
openai>=1.0.0
python-dotenv>=1.0.0
